services/observer/requirements.txt
websockets==12.0
python-dotenv==1.0.1